Added typescritp in project
