+++ src/App.tsx
@@ -1,7 +1,9 @@
+import React from 'react';
 import './App.css';
 import HomePage from './Components/HomePage';
 import Navbar from './Components/Navbar';
 import ProductState from './ProductsContext/ProductState';
+
 
 function App() {
   return (
@@ -13,5 +15,11 @@ function App() {
     </ProductState>
   );
 }
+
+
+
+
+
+
 
 export default App;
