+++ src/Components/HomePage.tsx
@@ -3,7 +3,7 @@ import productContext from '../ProductsContext/productContext'
 
 const HomePage = () => {
     const context = useContext(productContext)
-    const {products, getProducts } = context
+    const {products, getProducts }: any = context
 
     useEffect(() => {
         getProducts()
@@ -14,9 +14,9 @@ const HomePage = () => {
   return (
     <div className='container mt-3'>
         <div className='row'>
-        {products.map((product) => {
+        {products.map((product: any) => {
             return <>
-            <div className='col-md-3 my-3 border text-center'>
+            <div key={product.id} className='col-md-3 my-3 border text-center'>
                 <div className="card rounded-3 border border-black border-3" style={{width: "20rem"}} >
                     <img src={product.images[0]} className="card-img-top rounded-top" alt="..." style={{height: "300px"}}/>
                         <div className="card-body" >
