+++ src/ProductsContext/productContext.tsx
@@ -1,5 +1,5 @@
 import { createContext } from "react";
 
-const productContext = createContext()
+const productContext = createContext("")
 
 export default productContext;
